point-cloud-dynamics-model/requirements.txt
# Requirements for Point Cloud Dynamics Model

# Core dependencies
torch>=2.0.0
pyyaml>=6.0

# Optional but recommended for acceleration
flash-attn>=2.0.0  # Install with: pip install flash-attn --no-build-isolation

# For development and testing
numpy>=1.20.0
matplotlib>=3.5.0